backend/requirements.txt
Django
djangorestframework
django-cors-headers
djongo
pymongo==3.12.3
